Convert helpers package to ES modules with rollup cjs/umd builds
no issue
- move package source files into `src` and convert to ES modules
- add rollup config for node CJS build and browser UMD builds
- update `package.json`s `main` and `module` properties to point at CJS and ES builds respectively
  - `main` points at the CJS build, it's used when importing into a node project so it only converts from ES to CJS modules with no transpilation or bundling
  - `module` points directly at ES source code, when imported into projects with a build environment such as rollup, webpack, etc then the project's bundler will do it's thing

The UMD build bundles all dependencies (in this case a large chunk of lodash), transpiles to ES5 for older browser support, and minifies to reduce load size. This file should be copied or imported directly into projects that rely on plain concatenation of assets or have no build step.