Bumped @tryghost/admin-api-schema to 2.0.0-rc.4
no-issue

We would like to release @tryghost/admin-api-schema as version 2.0.0 but
lerna will only allow us to release packages which have had changes in
them since their last release. Well, this is a change. And it's solely
for the purpose of allowing us to release 2.0.0. This version is not
tagged, nor is it published to npm.