🐛 Fixed ReDoS susceptible regexs
fixes https://github.com/TryGhost/Team/issues/3544

The regexs for counting words and counting images were susceptible
to ReDoS and could be problematic in Ghost when the a post's HTML
contains a tag with a very large attribute (i.e an image src being
a base64 encoded image). This commit tweaks these regexs to make
them more efficient and circumvent the problem.