Removed `currentCountQuery` from limit service config for `customIntegrations`
ref BAE-331
ref https://ghost.slack.com/archives/C02G9E68C/p1750776240012249

Since the introduction of the `customIntegrations` limit, it was been treated as a `Flag` limit, which is a simple on/off limit. The query that was assigned to in the config has never been executed and is not found in use anywhere in our repos.
To be able to let customers use a feature, which they have enabled, after we impose a limit for that feature, we're exploring the option to use `currentCountQuery` to overwrite the imposed limit if the feature is already being used. Having a query in the config for custom integrations messes with that idea, besides the fact that it's a pointless query.