Refactored the absolute to transform ready helper
By using early returns we pull the core logic out of nested try/catch
and if blocks.

There is a slight functional difference here in the first return,
because we continue the flow if the url passed it fails parsing.

This is an edge-case that we don't hit in reality, and it's also still
safe, because the later checks will catch an failed parsing, so the
behaviour of the helper hasn't changed, despite a flow change.