Supported `__GHOST_URL__` transforms for CDN asset URLs
ref https://linear.app/ghost/issue/PRO-1524

In order to correctly store assets URLs when they are on a CDN we need
to support transforming more than just the siteUrl to `__GHOST_URL__`.

This implementation allows for custom asset urls for images, media & files
and supports these urls including a subdomain. The bulk of the logic is
in the `absoluteToTransformReady` & `transformReadyToAbsolute` utils
which are then used by all of the more specific transforms.

The html & markdown transforms have an early return that we needed
to modify to ensure still runs for CDN urls

The plaintext transform had a regex to optimise the replacements
which needed updating to match the CDN urls too.

We also export the static constants for consumption by Ghost, in-line
with the STATIC_IMAGE_PREFIX pattern we have.